+++ test/foundry/Swap.t.sol
@@ -63,14 +63,14 @@ contract TestSwap is Test {
 
     function test_swap_tokenA_liquidity_in_range_succeeds() public mintApproveTokens {
         // fetch slotIndex of pool
-        (, , , , , , UD60x18 sqrtPrice, int128 slotIndex, , , , ) = hyper.pools(poolId);
+        (, , , , , , UD60x18 sqrtPrice, int24 slotIndex, , , , ) = hyper.pools(poolId);
 
         emit log("before swap: sqrt price");
         emit log_uint(unwrapUD60x18(sqrtPrice));
 
         int256 liquidity = int256(1e18);
-        int128 lowerSlotIndex = slotIndex - 10;
-        int128 upperSlotIndex = slotIndex + 10;
+        int24 lowerSlotIndex = slotIndex - 10;
+        int24 upperSlotIndex = slotIndex + 10;
 
         // add liquidity around the pool's slot index
         hyper.updateLiquidity(poolId, lowerSlotIndex, upperSlotIndex, liquidity);
@@ -93,14 +93,14 @@ contract TestSwap is Test {
 
     function test_swap_tokenB_liquidity_in_range_succeeds() public mintApproveTokens {
         // fetch slotIndex of pool
-        (, , , , , , UD60x18 sqrtPrice, int128 slotIndex, , , , ) = hyper.pools(poolId);
+        (, , , , , , UD60x18 sqrtPrice, int24 slotIndex, , , , ) = hyper.pools(poolId);
 
         emit log("before swap: sqrt price");
         emit log_uint(unwrapUD60x18(sqrtPrice));
 
         int256 liquidity = int256(1e18);
-        int128 lowerSlotIndex = slotIndex - 10;
-        int128 upperSlotIndex = slotIndex + 10;
+        int24 lowerSlotIndex = slotIndex - 10;
+        int24 upperSlotIndex = slotIndex + 10;
 
         // add liquidity around the pool's slot index
         hyper.updateLiquidity(poolId, lowerSlotIndex, upperSlotIndex, liquidity);
